Treat "items" as special keyword for Array basic type
In json.encoder, the implementation of `_make_iterencode` relies on
the "items" to get all keys/values in dictionary(check the innter
function, `_iterencode_dict`). Therefore, defining the `items` as
field in dataclass can cause the function borken. This field will
override the attribute in the `dict` type, which cause the error

TypeError: 'str' object is not callable

thrown(dependes on value's type you passed to this field). To
prevent the problem, now we reserve the `items` as special keyword
field in apispec-helper object. All objects need to use this filed
should use the `items_` instead.